Extract VersionUtils class
